% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wk_api.R
\name{get_wk_data}
\alias{get_wk_data}
\title{Get all of the available data from the WaniKani API}
\usage{
get_wk_data(api_key, api_ver = API_VER, cache_dir = NULL,
  max_recent_unlocks = 100, critical_items_pct = 75, time_message = TRUE,
  print_url = TRUE)
}
\arguments{
\item{api_key}{the users API key}

\item{api_ver}{the API version to use}

\item{cache_dir}{directory read cached data from}

\item{time_message}{print a time message}

\item{print_url}{print the URL string being used, useful for debugging}

\item{param}{max_recent_unlocks max number of recent unlocks to get; an opt_arg}

\item{param}{critical_items_pct percentage threshold for critical items; an opt_arg}
}
\value{
a list of lists, one item for each "data unit"
}
\description{
Retrieves all of the data from the the WaniKani API
}
