% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wk_api.R
\name{get_wk_data_by_name}
\alias{get_wk_data_by_name}
\title{Get a specific "data unit" from the WaniKani API}
\usage{
get_wk_data_by_name(data_name, api_key, api_ver = API_VER, opt_arg = NULL,
  time_message = TRUE, print_url = TRUE)
}
\arguments{
\item{data_name}{the name of the "data unit" -- see the API docs}

\item{api_key}{the users API key}

\item{api_ver}{the version of the API to use; defaults to API_VER}

\item{opt_arg}{the optional URL argument that can be used for some "data units"}

\item{time_message}{print a time message}

\item{print_url}{print the URL string being used, useful for debugging}
}
\value{
results in the form of a list
}
\description{
Extracts a unit of data from the WaniKani API by name.
The names correspond to the strings used in the URL
requests and can be found on the WaniKani API documentation page.
}
